+++ modules/learn/learned.db
@@ -1,0 +1,1 @@
+{"ciao": ["miao", "bau"]}
