+++ modules/random/phrases/proverbs.txt
