+++ modules/random/phrases/theworld.txt
